+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Version 0.2.1-beta.1
-- Update app icon #31
 - Add confirmation dialog to remove environment #30
+- Update app icon #31
+- Show error dialog when remove fails #32
 
 ## Version 0.2.0-beta.1
 - Initial public beta release